SetWindowLong() arg type fix

+++ Ninjacrab.PersistentWindows.Solution/Common/WinApiBridge/User32.cs
@@ -424,7 +424,7 @@ namespace PersistentWindows.Common.WinApiBridge
         public static extern long GetWindowLong(IntPtr hWnd, int nIndex);
 
         [DllImport("user32.dll", SetLastError = true)]
-        public static extern int SetWindowLong(IntPtr hWnd, int nIndex, long dwNewLong);
+        public static extern int SetWindowLong(IntPtr hWnd, int nIndex, uint dwNewLong);
 
         public const int GWL_EXSTYLE = -20;
         public const long WS_EX_TOPMOST = 0x00000008L;
